refactor : 일부 validate 메서드의 위치를 뷰에서 도메인으로 변경
- 구매 금액이 로또 1장 가격으로 나누어 떨어지는지 검증하는 메서드
- 로또 번호, 보너스 번호의 범위가 1~45 사이인지 검증하는 메서드
- 입력한 로또 번호의 개수가 6개인지 검증하는 메서드
